Rearrange metadata module
